add FormEnter
