+++ src/components/CashFlow/CashFlow.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import { useState, useEffect } from 'react';
 import { FaTrashAlt } from 'react-icons/fa';
 import s from './CashFlow.module.css';
 
@@ -6,9 +6,22 @@ const classsTabRight = s.tab + ' ' + s.right_tab;
 const classsTabLeft = s.tab + ' ' + s.left_tab;
 
 const CashFlow = ({ arey }) => {
+  const [emptyArray, setEmptyArray] = useState([]);
+
+  useEffect(() => {
+    if (arey.length >= 9) return;
+    const arrayLength = 9 - arey.length;
+    const newArray = new Array(arrayLength).fill(5);
+
+    setEmptyArray(newArray);
+  }, [arey.length]);
+
+  console.log(`arey`, arey);
+  console.log(`emptyArray`, emptyArray);
+
   return (
     <div>
-      <table className={s.table23}>
+      <table className={`${s.table23} ${s.scrollbar}`}>
         <thead>
           <tr>
             <th className={classsTabLeft}>ДАТА</th>
@@ -18,7 +31,8 @@ const CashFlow = ({ arey }) => {
             <th className={classsTabRight}>УДАЛЕНИЕ</th>
           </tr>
         </thead>
-        <tbody>
+
+        <tbody className={s.scrollbar}>
           {!!arey.length &&
             arey.map(({ date, name, category, amount, id }) => (
               <tr key={id} className={s.field}>
@@ -33,6 +47,17 @@ const CashFlow = ({ arey }) => {
                 </td>
               </tr>
             ))}
+          {emptyArray.map((_, idx) => {
+            return (
+              <tr key={idx}>
+                <td className={s.emptyLine}></td>
+                <td className={s.emptyLine}></td>
+                <td className={s.emptyLine}></td>
+                <td className={s.emptyLine}></td>
+                <td className={s.emptyLine}></td>
+              </tr>
+            );
+          })}
         </tbody>
       </table>
     </div>
